put the Carousel in the layout template
