+++ packages/client/src/components/SimilarMoviesCarousel/Carousel.css
@@ -1,0 +1,11 @@
+.view-carousel {
+  height: 300px;
+  width: 1000px;
+}
+img {
+  width: 330px;
+  height: 250px;
+  border-radius: 112.5px;
+  /* filter: drop-shadow(20px 20px 40px rgba(239, 13, 13, 0.968)); */
+  margin-top: 4px;
+}
