--- packages/client/src/components/ViewPage/Carousel/Carousel.css
@@ -1,12 +1,0 @@
-.view-carousel {
-  height: 400px;
-  width: 605px;
-}
-img {
-  background-color: #a1b6ec;
-  width: 200px;
-  height: 125px;
-  filter: drop-shadow(20px 20px 40px rgba(0, 0, 0, 0.25));
-  border-radius: 112.5px;
-  margin-top: 4px;
-}
