--- packages/client/src/containers/ViewPage/ViewPage.jsx
@@ -1,11 +1,0 @@
-import React from 'react';
-
-import Carousel from '../../components/ViewPage/Carousel/Carousel';
-
-export default function ViewPage() {
-  return (
-    <div>
-      <Carousel />
-    </div>
-  );
-}
